Refactor financial data structs
