+++ src/apps/workbench/backtester/data/json_data_parser.h
@@ -2,7 +2,7 @@
 
 #include <string>
 
-#include "apps/workbench/core/common_structs.h"
+#include "common/financial_data_types.h"
 
 #include <vector>
